--- src/main/java/Command.java
@@ -1,5 +1,0 @@
-public abstract class Command {
-
-    abstract void execute(ToDo taskList, Ui ui, Storage storage) throws KermitException;
-    abstract boolean isExit();
-}
